Support initramfs-tools in addition to dracut
NOTE: This has not been re-tested under dracut.

initramfs-tools has a copy_exec shell function that works like
dracut-install to install an executable file and associated shared
libraries.  We can use that to replace inst() with an appropriate
version at runtime.

Sourcing in the entire hook-functions may be a bad idea, but it's how we
have to access copy_exec.  We could use a wrapper script to isolate it,
I suppose.

copy_exec always tries to locate binaries into /usr/bin (for a merged
/usr), so we have to correct that so /bin/sh and other symlinks can
resolve properly.

dracut-install looks in PATH to find xenstore-read (I suppose), but
copy_exec does not do lookups.  Use `which` to do the lookup ourselves.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>